==> Code/02c_recordfiles_masterdatamanagement.ipynb <==
import numpy as np
import pandas as pd
import json
import os
import pickle
import csv

#persistate master data
masterdata_path = 'masterdata/'

elo_frames = {}
elo_frames['low'] = {'lb':0,'ub': 1062, 'label': 'Low Elo'}
elo_frames['medium'] = {'lb': 1062,'ub': 1501, 'label': 'Medium Elo'}
elo_frames['high'] = {'lb': 1501,'ub': 3000, 'label': 'High Elo'}

output = open(masterdata_path + 'elo_frames.pkl' , 'wb')
pickle.dump(elo_frames, output)
output.close()       

#load masterdata
with open('masterdata/masterdata_dict.pkl', 'rb') as f:
        masterdata_dict = pickle.load(f)


#UniteMasterdataÄ
#read table
df = pd.read_csv('masterdata/unit_masterdata.csv',delimiter = ';')


temp_md_dict = df.to_dict('index')
unit_masterdata_dict = {}

#correct and rearange entries
for entry in temp_md_dict:
    for val in temp_md_dict[entry]:
        if isinstance( temp_md_dict[entry][val], str):
            temp_md_dict[entry][val] = temp_md_dict[entry][val].strip('\xa0').strip()
        
for entry in temp_md_dict:
    unit_masterdata_dict[temp_md_dict[entry]['Unit']]= temp_md_dict[entry]

#sort unit by types
    
unit_type_dict = {}
types = []

for entry in unit_masterdata_dict:
    if unit_masterdata_dict[entry]['Type'] not in types:
        types.append(unit_masterdata_dict[entry]['Type'])

#init lists
for type in types:
    unit_type_dict[type]=[]
        
for entry in unit_masterdata_dict:
    unit_type_dict[unit_masterdata_dict[entry]['Type']].append(entry)


#check for leftover values
for unit in masterdata_dict['Queue']:
    if unit not in unit_masterdata_dict:
        print(unit)

output = open('masterdata/unit_type_dict.pkl' , 'wb')
pickle.dump(unit_type_dict, output)
output.close()


output = open('masterdata/unit_masterdata_dict.pkl' , 'wb')
pickle.dump(unit_masterdata_dict, output)
output.close()


#Building Masterdata
#read table
df = pd.read_csv('masterdata/building_masterdata.csv',delimiter = ';')


temp_md_dict = df.to_dict('index')
building_masterdata_dict = {}

#correct and rearange entries
for entry in temp_md_dict:
    for val in temp_md_dict[entry]:
        if isinstance( temp_md_dict[entry][val], str):
            temp_md_dict[entry][val] = temp_md_dict[entry][val].strip('\xa0').strip()
        
for entry in temp_md_dict:
    building_masterdata_dict[temp_md_dict[entry]['Building']]= temp_md_dict[entry]

#sort unit by types
    
building_type_dict = {}
types = []

for entry in building_masterdata_dict:
    if building_masterdata_dict[entry]['Type'] not in types:
        types.append(building_masterdata_dict[entry]['Type'])

#init lists
for type in types:
    building_type_dict[type]=[]
        
for entry in building_masterdata_dict:
    building_type_dict[building_masterdata_dict[entry]['Type']].append(entry)
    
#check for leftover values
for building in masterdata_dict['Build']:
    if building not in building_masterdata_dict:
        print(building)

output = open('masterdata/building_type_dict.pkl' , 'wb')
pickle.dump(building_type_dict, output)
output.close()


output = open('masterdata/building_masterdata_dict.pkl' , 'wb')
pickle.dump(building_masterdata_dict, output)
output.close()


#Building Masterdata
#read table
df = pd.read_csv('masterdata/research_masterdata.csv',delimiter = ';')


temp_md_dict = df.to_dict('index')
research_masterdata_dict = {}

#correct and rearange entries
for entry in temp_md_dict:
    for val in temp_md_dict[entry]:
        if isinstance( temp_md_dict[entry][val], str):
            temp_md_dict[entry][val] = temp_md_dict[entry][val].strip('\xa0').strip()
        
for entry in temp_md_dict:
    research_masterdata_dict[temp_md_dict[entry]['Technology']]= temp_md_dict[entry]

#sort unit by types
    
research_type_dict = {}
types = []

for entry in research_masterdata_dict:
    if research_masterdata_dict[entry]['Type'] not in types:
        types.append(research_masterdata_dict[entry]['Type'])

#init lists
for type in types:
    research_type_dict[type]=[]
        
for entry in research_masterdata_dict:
    research_type_dict[research_masterdata_dict[entry]['Type']].append(entry)
    
#check for leftover values
for technology in masterdata_dict['Research']:
    if technology not in research_masterdata_dict:
        print(technology)

output = open('masterdata/research_type_dict.pkl' , 'wb')
pickle.dump(research_type_dict, output)
output.close()


output = open('masterdata/research_masterdata_dict.pkl' , 'wb')
pickle.dump(research_masterdata_dict, output)
output.close()

full_type_dict = {}

for u in unit_type_dict:
    full_type_dict[f'Unit {u}'] = unit_type_dict[u]
    
for b in building_type_dict:
    full_type_dict[f'Building {b}'] = building_type_dict[b]
    
for t in research_type_dict:
    full_type_dict[f'Technology {t}'] = research_type_dict[t]

inv_full_type_dict = {}

for c in full_type_dict:
    for e in full_type_dict[c]:
        inv_full_type_dict[e] = c

output = open('masterdata/full_type_dict.pkl' , 'wb')
pickle.dump(full_type_dict, output)
output.close()


output = open('masterdata/inv_full_type_dict.pkl' , 'wb')
pickle.dump(inv_full_type_dict, output)
output.close()

##### gaia masterdata
"""
redundant_gaia_objects = gaia_objects
relevant_gaia_objects = {}  
relevant_gaia_objects['tree'] = []

for item in gaia_objects:
    if "Tree" in item:
        relevant_gaia_objects['tree'].append(item)     
        
relevant_gaia_objects['animal_gatherable'] = ['Deer','Elephant','Ibex','Iron Boar','Javelina','Ostrich','Rhinoceros','Wild Boar','Zebra']
relevant_gaia_objects['animal_hostile'] = ['Bear','Crocodile','Dire Wolf','Jaguar', 'Komod Dragon','Lion','Rabid Wolf','Snow Leopard','Tiger','Wolf']
relevant_gaia_objects['animal_herdable'] = ['Cow A','Cow B','Cow C','Cow D','Goat','Goose','Llama','Pig','Sheep','Turkey','Water Buffalo']
relevant_gaia_objects['aquatic_animal'] = ['Shore Fish', 'Fish (Snapper)', 'Fish (Tuna)','Fish (Salmon)', 'Fish (Dorado)', 'Box Turtles']
relevant_gaia_objects['gold_mine']  = ['Gold Mine']
relevant_gaia_objects['stone_mine'] = ['Stone Mine']
relevant_gaia_objects['relic'] = ['Relic']
relevant_gaia_objects['food'] = ['Fruit Bush', 'Berry Bush', 'Forage Bush']
relevant_gaia_objects['p1_tc'] = ['p1_tc']
relevant_gaia_objects['p2_tc'] = ['p2_tc']
"""


import pickle

with open('masterdata/gaia_masterdata.pkl', 'rb') as f:
    gaia_md = pickle.load(f)

gaia_masterdata_reversed = {}

for ot in gaia_md:
    for item in gaia_md[ot]:
        gaia_masterdata_reversed[item] = ot 



output = open(masterdata_path + 'gaia_masterdata_reversed.pkl' , 'wb')
pickle.dump(gaia_masterdata_reversed, output)
output.close()       

### merge masterdata

with open(f'masterdata/unit_masterdata_dict.pkl', 'rb') as f:
        unit_masterdata_dict = pickle.load(f)
        
with open(f'masterdata/building_masterdata_dict.pkl', 'rb') as f:
        building_masterdata_dict = pickle.load(f)
        
with open(f'masterdata/research_masterdata_dict.pkl', 'rb') as f:
        research_masterdata_dict = pickle.load(f)




full_masterdata_dict = {}

input_md_type_dict = {'research': research_masterdata_dict,
                      'queue': unit_masterdata_dict,
                      'build': building_masterdata_dict}

resources = ['Food','Wood','Gold','Stone']

for input_type in input_md_type_dict:
    
    for key in input_md_type_dict[input_type]:
        full_masterdata_dict[key] = input_md_type_dict[input_type][key]
        full_masterdata_dict[key]['input_type'] = input_type
        
        #calc total resources
        total_resources = 0
        for subkey in full_masterdata_dict[key]:
            if subkey in resources:
                total_resources += full_masterdata_dict[key][subkey]
        
        full_masterdata_dict[key]['total_resources'] = total_resources

output = open('masterdata/full_resource_masterdata_dict.pkl' , 'wb')
pickle.dump(full_masterdata_dict, output)
output.close()
