==> Code/02d_recordfiles_filter.ipynb <==
import pandas as pd
import numpy as np
import pickle
import matplotlib.pyplot as plt

with open('data/scraped_matches/parsed_matches/final_matches.pkl', 'rb') as f:
    matches = pickle.load(f)



arabia_matches = {}

for match_id in matches:
    if matches[match_id]['map'] == 'Arabia':
        arabia_matches[match_id] = matches[match_id]
        
len(arabia_matches)

arabia_matches.head()

#final persitate matches
file_name = 'data/scraped_matches/parsed_matches/arabia_matches.pkl'
output = open(file_name, 'wb')
pickle.dump(arabia_matches, output)
output.close()
