==> Code/03e_recordfiles_input_cnn_preprocessing.ipynb <==
import numpy as np
import pandas as pd
import torch
import json
import os
import pickle
from collections import OrderedDict
import psutil
import gc

def print_txt(msg: str):
    print(msg, end = '\r')
    text_file = open("Outputs/input_cnn_preprocessing.txt", "w")
    text_file.write(msg)
    text_file.close()

##directorys
input_path = 'data/scraped_matches/inputs/'
gaia_path = 'data/scraped_matches/gaia_data/'

#players
player_ids = ['p1','p2']
player_winner_dict = {0: 'p1',1: 'p2'}

#load data
with open('data/scraped_matches/parsed_matches/arabia_matches.pkl', 'rb') as f:
    org_matches = pickle.load(f)
matches = org_matches

#load masterdata
with open('masterdata/masterdata_dict.pkl', 'rb') as f:
        masterdata_dict = pickle.load(f)

with open('masterdata/unit_type_dict.pkl', 'rb') as f:
        unit_type_dict = pickle.load(f)        

with open('masterdata/unit_masterdata_dict.pkl', 'rb') as f:
        unit_masterdata_dict = pickle.load(f)      
        
        


match = matches['50453403']
match_id = '50453403'
match

with open(matches[match_id]['input_fn'], 'rb') as f:
            inputs = pickle.load(f)

type_dict = {}
for id in inputs:
    input = inputs[id]
    if input['type'] not in type_dict:
        type_dict[input['type']] = input


type_dict.keys()
#start with move_target_build
#inputs

rel_input_types = ['Build','Move','Target']
timelimit = 1800

input_matchids = []
input_labels = {}

input_level_types = {}

for index,match_id in enumerate(matches):
    #determine winner of each match
    winner = player_ids[matches[match_id]['winner']]
    
    #open inputs
    with open(matches[match_id]['input_fn'], 'rb') as f:
                    inputs = pickle.load(f)
    
    """
    for p_id in player_ids:
        create 2 views per match
    view_id = str(match_id) + '_' + p_id

    if player_winner_dict[match['winner']] == p_id:
        win_label = 1
    else:
        win_label = 0
    
    
    """
    array_list = []
    #init empty arrays of shape map size
    for p_id in player_ids:  
        arrays = {}
        for type in rel_input_types:
            arrays[type] = np.zeros(shape=(120,120))

        for input_id in inputs:
            input = inputs[input_id]


            if input['type'] in rel_input_types and input['ts_seconds']<=timelimit and p_id == input['player']:
                try:
                    #fill values into arrays
                    arrays[input['type']][int(input['y_pos'])-1][int(input['x_pos'])-1] = 1
                except:
                    print(input)


        
        for i in arrays:
            array_list.append(arrays[i])

    md_array =  np.stack(array_list, axis=0)
    tensor = torch.tensor(md_array)

    #persistate tensor
    output = open(f'data/feature_maps/inputs/{match_id}.pkl' , 'wb')
    pickle.dump(tensor, output)
    output.close()

    #safe label and match_id
    input_labels[match_id] = match['winner']#win_label
    input_matchids.append(match_id)

    #input_labels[view_id] = win_label
    #input_matchids.append(view_id)

    if index % 1000 == 0:
        print_txt(f'{index}/{len(matches)} - {tensor.shape} ')



output = open('data/feature_maps/input_matchids.pkl' , 'wb')
pickle.dump(input_matchids, output)
output.close()
output = open('data/feature_maps/input_labels.pkl' , 'wb')
pickle.dump(input_labels, output)
output.close()


print_txt("check")
