==> Code/04_recordfile_datavalidation.ipynb <==
ls

import pickle
import numpy as np
import pandas as pd
from pathlib import Path


#load tab_data
with open('data/minn/recordfile_tab_data_raw.pkl' , 'rb') as f:
        tab_data_df = pickle.load(f)
tab_data_df.shape

tab_data_df.head()

#check for all imgf if file exists
found = 0
not_found = 0
index_tbd = []


for index, row in tab_data_df.iterrows():
    fn = row['imgf_Move_layer']
    if Path(f'data/minn/{fn}').is_file():
        found += 1
    else:
        not_found += 1
        index_tbd.append(index)
        

found, not_found

sync_tab_data_df = tab_data_df.drop(index_tbd)

sync_tab_data_df.shape

#write back
output = open('data/minn/recordfile_tab_data.pkl' , 'wb')
pickle.dump(sync_tab_data_df, output)
output.close()
